refit: shells by the box, a drive you can lose, and rounds you pay for
The ammo hardpoint was authored and unused. It now takes one of two boxes --
CANNON SHELL BOX or MISSILE CONTAINER -- built as a single AmmoStore kind with
an AmmoPool discriminator, the same shape the two shield emitters already have,
which is where the swap-replaces-rather-than-stacks behaviour comes from. One
slot, so carrying spares is a choice about which weapon this pilot expects to
run dry. Neither is researched (`research = false`): shells are not a
technology, so they never appear on the permanent board and no unlocked rank
gates them.

ENGINE fills the engine slot and OVERBURN moves to a thruster marker of its
own -- one category per non-mounted fitting, or the board cannot tell which of
them sits where. Rank I is issued rather than researched: it costs no Tech, so
every faction starts holding it (IssueFreeUnlocks) and every hull spawns with
it fitted. It scales nothing, because rank I *is* the motor the hull was drawn
with and a hull's authored [physics] numbers are what everything else is tuned
against; the ranks above it compound. Pull the drive at a yard and the hull
steers, coasts and falls but cannot add a metre per second of its own. A
freighter carries no loadout at all and is not a fighter with its engine
pulled, so it keeps flying on its hull's own numbers.

Landing no longer reloads anything for free. Rounds are a purchase now:
RESUPPLY prices a refill at 75% of what the hardware feeding that magazine
cost to fit, pro-rata for the fraction actually missing, and the server prices
it again on arrival so a click cannot ask for a discount. It buys nothing for a
magazine whose last mount has been pulled. The installed list shows what each
fitting has left, red at zero.

Prerequisites: heavy cannons behind the light guns, the overburn behind a
drive, plating behind the bubble. That last one cannot be asked of the hull --
the two emitters replace each other, so no hull ever holds a bubble at the
moment it fits plating -- so a prerequisite that is mutually exclusive with
what it gates is checked against what the faction has learned instead.

Ammo counts widen to uint16 through the component, the wire and the HUD, so a
deep magazine plus a full box is not bounded by 255. Protocol 8.

The tech window gains a close button in the titlebar and another in the footer,
and -- the reason the trigger used to cost frames -- its refresh now skips a
window nobody is looking at. Live round counts stay out of TechNodeView's
equality operator and update in place; leaving them in rebuilt the entire board
once per round fired. The counters and the refit denial stay outside that gate:
one is on the HUD, the other writes to the chat log on its way out.

+++ include/gravitaris/game/component/ship-loadout.hpp
@@ -36,10 +36,10 @@ struct ShipLoadout {
     // it through, not a second rack.
     std::uint8_t missileBays = 0;
 
-    std::uint8_t missileAmmo = 0;
+    std::uint16_t missileAmmo = 0;
     // Rounds left in the cannon's magazine. Deeper than the rack by an order
     // of magnitude -- see the HUD, which draws one tick per ten.
-    std::uint8_t cannonAmmo = 0;
+    std::uint16_t cannonAmmo = 0;
     UpgradeLevels levels;
     // Shield charge, 0..ShipStats::shieldCapacity. ShieldSystem refills it;
     // DamageSystem spends it ahead of the hull. For plating this is the sum of
